insights is now included as a module with post, for added integration
