+++ app/controllers/posts_controller.rb
@@ -11,6 +11,6 @@ class PostsController < ApplicationController
 
   def show
     post = Post.find(params[:id])
-    render json: relevant_data
+    render json: post.to_json
   end
 end
